moved screen to pages directory
